Title: 실수 입력받아 계산, Time: 47ms, Memory: 16MB, Status: Passed - Codetree

+++ 250422/실수 입력받아 계산/calculation-by-inputting-a-real-number.py
@@ -1,3 +1,3 @@
-a = int(input())
-b = int(input())
+a = float(input())
+b = float(input())
 print(f'{a+b:.2f}')
